rename
https://coveord.atlassian.net/browse/CDX-465
